rename sourceFile to source-file to be consistent
