+++ src/consumer/component/sources/index.js
@@ -3,7 +3,7 @@ import Specs from './specs';
 import Source from './source';
 import Dist from './dist';
 import License from './license';
-import SourceFile from './sourceFile';
+import SourceFile from './source-file';
 
 export {
   Impl,
